pyproject.toml
[project]
name = "fastapi-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.17.0",
    "apscheduler==3.11.0",
    "bcrypt==4.1.2",
    "boto3==1.38.27",
    "botocore==1.38.46",
    "email-validator==2.3.0",
    "factory-boy==3.3.3",
    "fastapi[standard]==0.120.0",
    "fastapi-mail==1.5.0",
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-logging>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "passlib==1.7.4",
    "pydantic==2.12.3",
    "pydantic-settings==2.11.0",
    "pyright>=1.1.407",
    "pytest==8.4.2",
    "python-dotenv>=1.2.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "requests==2.32.3",
    "ruff==0.14.2",
    "sqlalchemy==2.0.44",
    "uvicorn==0.38.0",
    "slowapi>=0.1.9",
]

[tool.setuptools.packages.find]
include = ["app"]
exclude = ["migrations", "tests"]
